+++ src/main/java/blusunrize/immersiveengineering/client/render/ItemRenderRevolver.java
@@ -1,81 +1,87 @@
 package blusunrize.immersiveengineering.client.render;
 
+import net.minecraft.client.renderer.OpenGlHelper;
 import net.minecraft.client.renderer.entity.RenderItem;
 import net.minecraft.item.ItemStack;
 import net.minecraftforge.client.IItemRenderer;
+import net.minecraftforge.client.model.obj.WavefrontObject;
 
 import org.lwjgl.opengl.GL11;
 
 import blusunrize.immersiveengineering.client.ClientUtils;
 import blusunrize.immersiveengineering.client.models.ModelRevolverVariable;
 import blusunrize.immersiveengineering.common.items.ItemRevolver;
-import blusunrize.immersiveengineering.common.util.ItemNBTHelper;
 
 public class ItemRenderRevolver implements IItemRenderer
 {
 	static ModelRevolverVariable model = new ModelRevolverVariable();
+	static WavefrontObject modelobj = ClientUtils.getModel("immersiveengineering:models/revolver.obj");
 
 	@Override
 	public boolean handleRenderType(ItemStack item, ItemRenderType type)
 	{
-		return item.getItemDamage()!=2;
+		return item.getItemDamage()!=1;
 	}
 	@Override
 	public boolean shouldUseRenderHelper(ItemRenderType type, ItemStack item, ItemRendererHelper helper)
 	{
-		return item.getItemDamage()!=2;
+		return item.getItemDamage()!=1;
 	}
 
 	@Override
 	public void renderItem(ItemRenderType type, ItemStack item, Object... data)
 	{
 		GL11.glPushMatrix();
-		GL11.glRotatef(180, 0, 0, 1);
 		if(type==ItemRenderType.EQUIPPED_FIRST_PERSON)
-			GL11.glRotatef(45, 0, 1, 0);
+			GL11.glRotatef(-135, 0, 1, 0);
 		else
-			GL11.glRotatef(-45, 0, 1, 0);
+			GL11.glRotatef(45, 0, 1, 0);
 
 		if(type==ItemRenderType.EQUIPPED_FIRST_PERSON)
 		{
-			GL11.glTranslatef(-1.25f,-.75f,-.75f);
-			GL11.glScalef(1, 1, .75f);
+			GL11.glTranslatef(-1.15f,.5f,-1f);
+			GL11.glScalef(.625f, .75f, .75f);
 		}
 		else if(type==ItemRenderType.EQUIPPED)
 		{
-			GL11.glRotatef(65, 1, 0, 0);
-			GL11.glTranslatef(0,.5625f,.7f);
-			GL11.glScalef(.5f,.5f,.5f);
+			GL11.glRotatef(-90, 0, 1, 0);
+			GL11.glRotatef(65, 0, 0, 1);
+			GL11.glTranslatef(.3f,-.7f,0f);
+			GL11.glScalef(.375f,.375f,.375f);
 		}
 		else if(type==ItemRenderType.ENTITY)
 		{
-			if(!RenderItem.renderInFrame)
-				GL11.glTranslatef(0,-.2f,0);
+			float scale = .25f;
 			GL11.glRotatef(45, 0, 1, 0);
-			GL11.glRotatef(-45, 1, 0, 0);
-			GL11.glTranslatef(0,-.1f,0.2f);
-			GL11.glScalef(.25f,.25f,.25f);
+			if(RenderItem.renderInFrame)
+			{
+				GL11.glRotatef(20, 0, 0, 1);
+				scale = .16f;
+				GL11.glTranslatef(.05f,.1f,0);
+			}
+			GL11.glScalef(scale,scale, scale);
 		}
 		else if(type==ItemRenderType.INVENTORY)
 		{
-			GL11.glRotatef(-90, 0, 1, 0);
-			GL11.glRotatef(-45, 1, 0, 0);
-			GL11.glTranslatef(0,-.1f,.3f);
-			GL11.glScalef(.4375f,.4375f,.4375f);
-			GL11.glEnable(3042);
+			GL11.glRotatef(0, 0, 1, 0);
+			GL11.glRotatef(-40, 0, 0, 1);
+			GL11.glTranslatef(0,.2f,.4f);
+			GL11.glScalef(.3f,.3f,.3f);
 		}
-		
+
 		ClientUtils.bindTexture(((ItemRevolver)item.getItem()).getRevolverTexture(item));
+		//		boolean b = (ItemNBTHelper.getInt(item, "upgrade")&1)==1;
+		//		model.Cylinder.isHidden = b;
+		//		model.CylinderAuto.isHidden = !b;
+		//		model.CylinderAuto_1.isHidden = !b;
+		//		model.CylinderAuto_2.isHidden = !b;
+		//		
+		//		model.render(null, 0, 0, 0, 0, 0, .0625f);
 		GL11.glColor4f(1, 1, 1, 1);
-		boolean b = (ItemNBTHelper.getInt(item, "upgrade")&1)==1;
-		model.Cylinder.isHidden = b;
-		model.CylinderAuto.isHidden = !b;
-		model.CylinderAuto_1.isHidden = !b;
-		model.CylinderAuto_2.isHidden = !b;
-		
-		model.render(null, 0, 0, 0, 0, 0, .0625f);
-		if(type==ItemRenderType.INVENTORY)
-			GL11.glDisable(3042);
+		GL11.glEnable(3042);
+		OpenGlHelper.glBlendFunc(770, 771, 0, 1);
+		modelobj.renderOnly(((ItemRevolver)item.getItem()).compileRender(item));
+		GL11.glDisable(3042);
 		GL11.glPopMatrix();
 	}